+++ engine/apps/alerts/models/alert_group.py
@@ -401,15 +401,34 @@ class AlertGroup(AlertGroupSlackRenderingMixin, EscalationSnapshotMixin, models.
         raise NotImplementedError
 
     @property
-    def permalink(self):
+    def slack_permalink(self):
         if self.slack_message is not None:
             return self.slack_message.permalink
 
     @property
+    def telegram_permalink(self) -> typing.Optional[str]:
+        """
+        This property will attempt to access an attribute, `prefetched_telegram_messages`, representing a list of
+        prefetched telegram messages. If this attribute does not exist, it falls back to performing a query.
+
+        See `apps.public_api.serializers.incidents.IncidentSerializer.PREFETCH_RELATED` as an example.
+        """
+        from apps.telegram.models.message import TelegramMessage
+
+        if hasattr(self, "prefetched_telegram_messages"):
+            return self.prefetched_telegram_messages[0].link if self.prefetched_telegram_messages else None
+
+        main_telegram_message = self.telegram_messages.filter(
+            chat_id__startswith="-", message_type=TelegramMessage.ALERT_GROUP_MESSAGE
+        ).first()
+
+        return main_telegram_message.link if main_telegram_message else None
+
+    @property
     def permalinks(self) -> Permalinks:
-        # TODO: refactor 'permalink' property (maybe 'slack_permalink'?) once we add the next permalink
         return {
-            "slack": self.permalink,
+            "slack": self.slack_permalink,
+            "telegram": self.telegram_permalink,
         }
 
     @property