+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
-## v1.0.37 (2022-09-23)
+## v1.0.38 (2022-09-30)
+
+- Fix exception handling for adding resolution notes when slack and oncall users are out of sync.
+- Fix all day events showing as having gaps in slack notifications
+- Improve plugin configuration error message readability
+- Add `telegram` key to `permalinks` property in `AlertGroup` public API response schema
+
+## v1.0.37 (2022-09-21)
 
 - Improve API token creation form
 - Fix alert group bulk action bugs